pyproject.toml
[project]
name = "metervision"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Ansari Ehteesham Aqeel", email = "an.ehteesham@gmail.com" }
]
requires-python = ">=3.12"
dependencies = [
    "dotenv>=0.9.9",
    "ipykernel>=6.30.1",
    "numpy==2.3.2",
    "python-box>=7.3.2",
    "pyyaml>=6.0.2",
    "streamlit>=1.47.1",
    "ultralytics>=8.3.174",
]

[build-system]
requires = ["uv_build>=0.8.3,<0.9.0"]
build-backend = "uv_build"
